+++ client/app/about/components/about.component.ts
@@ -2,8 +2,8 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-about',
-  templateUrl: './about.component.html',
-  styleUrls: ['./about.component.scss']
+  templateUrl: '../templates/about.component.html',
+  styleUrls: ['../styles/about.component.scss']
 })
 export class AboutComponent {
