+++ client/app/app-routing.module.ts
@@ -1,11 +1,11 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { AppComponent } from './app.component';
-import { HomeComponent } from './home/home.component';
-import { AboutComponent } from './about/about.component';
+import { AppComponent } from './core/components/main/app.component';
+import { CatsComponent } from './cats/components/cats.component';
+import { AboutComponent } from './about/components/about.component';
 
 const routes: Routes = [
-  { path: '', component: HomeComponent },
+  { path: '', component: CatsComponent },
   { path: 'about', component: AboutComponent }
 ];
