+++ client/app/cats/components/cats.component.spec.ts
@@ -1,7 +1,7 @@
 /* tslint:disable:no-unused-variable */
 
 import { TestBed, async } from '@angular/core/testing';
-import { HomeComponent } from './home.component';
+import { CatsComponent } from './cats.component';
 
 /*describe('Component: Home', () => {
   it('should create an instance', () => {
