+++ client/app/cats/components/cats.component.ts
@@ -2,15 +2,15 @@ import { Component, OnInit } from '@angular/core';
 import { Http } from '@angular/http';
 import { FormGroup, FormControl, Validators, FormBuilder } from '@angular/forms';
 
-import { DataService } from '../services/data.service';
-import { ToastComponent } from '../shared/toast/toast.component';
+import { CatsDataService } from '../services/cats.data.service';
+import { ToastComponent } from '../../shared/components/toast/toast.component';
 
 @Component({
   selector: 'app-home',
-  templateUrl: './home.component.html',
-  styleUrls: ['./home.component.scss']
+  templateUrl: '../templates/cats.component.html',
+  styleUrls: ['../styles/cats.component.scss']
 })
-export class HomeComponent implements OnInit {
+export class CatsComponent implements OnInit {
 
   cats = [];
   isLoading = true;
@@ -24,7 +24,7 @@ export class HomeComponent implements OnInit {
   weight = new FormControl('', Validators.required);
 
   constructor(private http: Http,
-              private dataService: DataService,
+              private dataService: CatsDataService,
               public toast: ToastComponent,
               private formBuilder: FormBuilder) { }
