+++ client/app/cats/services/cats.data.service.spec.ts
@@ -1,7 +1,7 @@
 /* tslint:disable:no-unused-variable */
 
 import { TestBed, async, inject } from '@angular/core/testing';
-import { DataService } from './data.service';
+import { CatsDataService } from './cats.data.service';
 
 /*describe('Service: Data', () => {
   beforeEach(() => {
