+++ client/app/cats/styles/cats.component.scss
