+++ client/app/core/components/main/app.component.spec.ts
