+++ client/app/core/components/main/app.component.ts
