+++ client/app/shared/components/toast/toast.component.scss
