+++ client/app/shared/components/toast/toast.component.ts
