+++ server/cats/controllers/cats.server.controller.ts
@@ -1,0 +1,8 @@
+import Cat from '../models/cats.server.model';
+import BaseCtrl from '../../base.server.controller';
+
+export default class CatsCtrl extends BaseCtrl {
+  model = Cat;
+}
+
+
