--- server/controllers/cats.ts
@@ -1,8 +1,0 @@
-import Cat from '../models/cat.model';
-import BaseCtrl from './base';
-
-export default class CatsCtrl extends BaseCtrl {
-  model = Cat;
-}
-
-
